Add user colors to mobile
Also use new models
