+++ backend/profiles/serializers/simple_user_serializer.py
@@ -7,4 +7,4 @@ class SimpleUserSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = User
-        fields = ['id', 'username']
+        fields = ['id', 'username', 'avatar_color']
