+++ backend/profiles/serializers/user_serializer.py
@@ -12,7 +12,7 @@ class UserSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = User
-        fields = ['id', 'username', 'watchlist', 'friends']
+        fields = ['id', 'username', 'avatar_color', 'watchlist', 'friends']
 
     def get_friends(self, obj):
         return SimpleUserSerializer(Friendship.objects.get_friends(obj), many=True).data
